initial camera, with viewport pixel positionn calculation
Given a resolution, fov and phi/theta rotations, the world position of the bottom-left corner
of the viewport is calculated, and a vector is calculated to shift a column or row
from it. debugPrintCorners() prints out the corners of the viewport in
the world in a format that can be copy-pasted into geogebra to validate.