map: add tiled floor generator, and map
gen_tile.go (written in go, evidently) takes some arguments and
generates a checkered floor, much like that in the famous whitted paper
render. This has been used in maps/checker.map, which is just 3sphere
with a floor. Rendering if SO much slower hence it isn't in the default
map.