"container" (quad) optimization structure
Shapes are now contained within a "ContainerQuad", which themselves can
be found within "Shape" structs.
```
container a <vec> b <vec> c <vec> d <vec> {
    <shapes>
}
```
Means that collisions will be tested on the quad abcd (a.k.a tris abc
and cab), and then only on its children if a collision is detected. used
in checker.map. Disable/Enable with "Use container quad optimisation".

The new object storage has issues that (likely) cause a SIGSEGV when
reloading the map.