render: fix accidentally sided triangles
When a triangle was axis aligned, the normal only has one non-zero
element, and if this was negative (i.e. you're looking from the "wrong" side), it would be mess up the projection and result in no "collision" with the tri. normal given to meetsTriangle() is now std::abs'ed element wise so axis aligned triangles show on both sides.