once again, fix SAH, + stopping condition
my understanding was wrong AGAIN, the surface area in question is not
that of the leaf nodes summed, but that of the AABB node containing
them! this makes leaps and bounds more sense, as made obvious when
trying to check the cost of the parent node. Speaking of, recursion now
terminates after reaching the set depth, OR when the calculated cost of
a split is greater than just traversing the parent.