rendering rework, separate window and render sizes
The width and height inputs no longer control the window size. Resizing
should be done by the user. Initially, the render dimensions are set to
that of the window, but will not follow if the latter changes. The
position of the quad we render to is calculated dynamically instead of
being constant, in order to letterbox the image rather than stretching
it. Additionally, a "Fit to window" button is added below the inputs to
se the render size to that of the window.