add material definition and reuse
materials are added to a material store, and can be defined per object
as before (sphere ... color #ffffff will make a unique material for the
object), or defined globally with "material <name> <properties>", then
used on an object like "sphere ... material <name>". Unique material
properties cannot be given to an object which uses a material, it will
cause a crash. The "edit materials" tab allows editing materials
directly, rather than through an object.