support un-baked transforms
transforms can now also be applied to general objects in the same way as
"include"s, however they are not baked into the object parameters and
are instead applied during ray traversal. They can be modified from the
UI too. Obviously breaks acceleration structures, a warning for which
has been added when enabling.