shape: transform application duplicates object
realised storing a transformed value and its original version in the
same object would break multithreading, so moved back to the old
approach of duplicating the object.

bakeTransform() applies a transform
to an object and resets the transform parameters, permanently "baking"
it.

applyTransform() makes a copy of the Shape, bakes the copy, then
returns it.