fix sphere normals within sphere, AAB uv customization, add specular control
AAB uvs can be tiled as triangles can, with an additional "face"
argument which determines which face to base the tiling on. Texture
scale is kept consistent across all faces, even those with differing
dimensions. Example: texture.tga:a,2,z will tile the texture twice
vertically on the z-facing pair of faces, and maintain aspect ratio.